`bigint` is an exact integer, and not a number that looks like one
The largest single blocker left, at 47 refusals, and it was hiding inside the
`Map` row until the last commit named what those blocked on.

# The width, and why it is visible

A `bigint` is arbitrary precision in the specification and 128 bits here. That
is a real difference, so it is drawn where it can be seen rather than felt: a
literal too large is refused at the place it is written, with its own digits in
the message.

The trade is measured, not assumed. Every `bigint` in the node profile is a
64-bit quantity -- `readBigUInt64BE`, an hrtime timestamp,
`0xffffffffffffffffn` -- and the comment above the first says why: "`bigint`
rather than `number`, because a `double` holds only 53 bits". A true bignum
puts a heap allocation into each of those, which is the opposite of what they
are reaching for. 128 bits covers all of them with 63 to spare, at the cost of
one C type and no allocation at all. What replaces it, when something needs
`2n ** 200n`, is the representation engines use: a small-integer fast path
beside a heap bignum, which is strictly more work than either half.

# Why it is its own `HirType`

It was `Int { bits: 128 }` for an afternoon, and the differential took it apart.

    let x = 1n; x <<= 40n; x |= 0xffn; Number(x >> 8n)

came back as 1 where node says 2^32. Constant folding had computed it -- and
correctly, by its own lights: a *number*'s shift masks its count to five bits,
so `1 << 40` is 256, `| 255` is 511, `>> 8` is 1. `**` reached
`nts_math_pow(2.0, 63.0)`, and the literal arrived as a `double`. Every pass
that knows the number rules had applied them, because nothing distinguished a
`bigint` from a number the specializer had proved integral.

Giving it a type turned six of those into compile errors at the match that had
to decide, which is the only way a second numeric semantics arrives without a
search for every pass that assumed the first. Two of the six were real
decisions rather than paperwork: an `async` function settling with one is
refused, because a promise's payload is a two-slot union of a double and a
pointer and keeping the low 53 bits of a `bigint` is precisely the loss it
exists to avoid; and it does not cross the N-API boundary, where node's
`bigint` is arbitrary precision and this one is not.

`ConstInt` widened from `i64` to `i128` for the same reason: `0xffffffffffffffffn`
is 2^64-1 and not an `i64`. Every narrower width still fits.

47 refusals to 11. The profile lowers 620 functions where it lowered 607, and
still verifies whole. What remains is `BigInt(x)` as a global, the typed-array
64-bit accessors, and reading one back out of an erased value.

`examples/bigint` is 319 cases: the literal forms, 2^64-1, truncating division
including a negative dividend, `**`, shifts past 64 bits, the bitwise
operators, and `Number()` being lossy above 2^53 in node and here alike.

+++ compiler/core/src/hir/simplify.rs
@@ -106,7 +106,7 @@ fn identity(func: &Func, value: ValueId) -> Option<ValueId> {
 }
 
 /// A value's integer constant, if it is one.
-fn constant(func: &Func, value: ValueId) -> Option<i64> {
+fn constant(func: &Func, value: ValueId) -> Option<i128> {
     match func.values[value.0 as usize].kind {
         OpKind::ConstInt(literal) => Some(literal),
         _ => None,
